requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==0.21.0
boto3==1.26.50
azure-servicebus==7.12.1
azure-data-tables==12.5.0